# docs/06-api/openapi-spec.yaml
openapi: 3.0.3
info:
  title: SupremeAI 2.0 API Specification
  description: OpenAPI specifications for SupremeAI 2.0 endpoints, including Email Service and GitHub Integration.
  version: 2.0.0
paths:
  /auth/gmail:
    post:
      summary: "Connect Gmail via OAuth"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                provider:
                  type: string
                  example: "gmail"
                scopes:
                  type: array
                  items:
                    type: string
                  example: ["https://www.googleapis.com/auth/gmail.modify"]
      responses:
        '200':
          description: Gmail connected successfully.
  /auth/imap:
    post:
      summary: "Connect generic IMAP"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                host:
                  type: string
                  example: "imap.gmail.com"
                port:
                  type: integer
                  example: 993
                username:
                  type: string
                  example: "supremeai@paykaribazar.com"
                app_password:
                  type: string
                  example: "app_password_secret"
      responses:
        '200':
          description: IMAP credentials connected.
  /github/connect:
    post:
      summary: "Connect customer GitHub repo"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                installation_id:
                  type: string
                  example: "from_github_app"
                repo_owner:
                  type: string
                  example: "customer-username"
                repo_name:
                  type: string
                  example: "customer-project"
      responses:
        '200':
          description: GitHub repository connected successfully.
  /github/improve:
    post:
      summary: "AI analyzes and improves code"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                repo:
                  type: string
                  example: "owner/repo"
                branch:
                  type: string
                  example: "main"
                improvement_type:
                  type: string
                  enum: [refactor, security, performance, docs]
                  example: "refactor"
      responses:
        '200':
          description: Analysis report and improvement recommendations generated.
  /github/push:
    post:
      summary: "Push improvements as PR"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                branch:
                  type: string
                  example: "supremeai-improvements-1718952000"
                commit_message:
                  type: string
                  example: "AI: Optimized database queries"
                files_changed:
                  type: array
                  items:
                    type: string
                  example: ["src/db.py", "src/cache.py"]
      responses:
        '200':
          description: Pull request generated successfully.
  /marketplace/search:
    post:
      summary: "Search across all marketplaces"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                query:
                  type: string
                  example: "PDF parser library"
                categories:
                  type: array
                  items:
                    type: string
                  example: ["npm", "pypi", "docker"]
                filters:
                  type: object
                  properties:
                    min_stars:
                      type: integer
                      example: 100
                    license:
                      type: array
                      items:
                        type: string
                      example: ["MIT", "Apache-2.0"]
                    last_updated:
                      type: string
                      example: ">6 months"
      responses:
        '200':
          description: Search results returned successfully.
  /marketplace/install:
    post:
      summary: "Auto-install discovered tool"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tool_id:
                  type: string
                  example: "npm:pdf-parse"
                target_environment:
                  type: string
                  example: "supremeai-worker-01"
                sandbox:
                  type: boolean
                  example: true
      responses:
        '200':
          description: Tool installed successfully.
  /github/discover:
    post:
      summary: "Find suitable GitHub repos"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                requirement:
                  type: string
                  example: "React component library for data tables"
                tech_stack:
                  type: array
                  items:
                    type: string
                  example: ["React", "TypeScript"]
                criteria:
                  type: object
                  properties:
                    min_stars:
                      type: integer
                      example: 500
                    min_contributors:
                      type: integer
                      example: 5
                    issues_open_ratio:
                      type: string
                      example: "<0.1"
                    last_commit:
                      type: string
                      example: "<3 months"
                    license:
                      type: array
                      items:
                        type: string
                      example: ["MIT", "Apache-2.0"]
                    has_typescript_defs:
                      type: boolean
                      example: true
      responses:
        '200':
          description: Discovered repositories returned.
  /github/implement:
    post:
      summary: "Integrate discovered repo into project"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                repo_url:
                  type: string
                  example: "https://github.com/TanStack/table"
                integration_method:
                  type: string
                  enum: ["npm", "submodule", "fork", "copy"]
                  example: "npm"
                target_project:
                  type: string
                  example: "customer-ecommerce-app"
      responses:
        '200':
          description: Repository integrated successfully.